Move blocks functions to wordpress sub-folder
